Use uint32 boot ids for time snapshots

+++ src/network/Esp32BaseNtp.h
@@ -21,7 +21,7 @@ public:
         bool synced;
         uint32_t epochSec;
         uint32_t uptimeSec;
-        uint16_t bootId;
+        uint32_t bootId;
         uint32_t bootStartEpochSec;
     };
 
@@ -35,9 +35,9 @@ public:
     static uint32_t timestamp();
     static TimeSnapshot snapshot();
     static void onTimeSynced(TimeSyncCallback callback);
-    static bool isCurrentBootEvent(uint16_t bootId);
-    static bool canResolveCurrentBootEvent(uint16_t bootId);
-    static bool resolveCurrentBootEvent(uint16_t bootId, uint32_t uptimeSec, uint32_t* epochSec);
+    static bool isCurrentBootEvent(uint32_t bootId);
+    static bool canResolveCurrentBootEvent(uint32_t bootId);
+    static bool resolveCurrentBootEvent(uint32_t bootId, uint32_t uptimeSec, uint32_t* epochSec);
     static void setServers(const char* s1, const char* s2 = nullptr, const char* s3 = nullptr);
     static bool formatTime(char* out, size_t len, const char* fmt);
     static const char* logTimeString();
